# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vector Client API
  description: API for managing endpoints, schemas (knowledge base), scrape sources, native AI agents, tokens, and logs.
  version: 1.1.0
servers:
  - url: http://localhost:3001
    description: Default local server
paths:
...
  /api/endpoints/{id}/schemas:
    post:
      summary: Add schemas (knowledge base) to endpoint
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                schemaIds: { type: array, items: { type: string, format: uuid } }
      responses:
        '200':
          description: Schemas associated
        '404':
          description: Endpoint not found
  /api/endpoints/{id}/schemas/{schemaId}:
    delete:
      summary: Remove schema from endpoint
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: schemaId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Schema removed
        '404':
          description: Not found
...
  /api/schemas:
    get:
      summary: List schemas (knowledge base)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of schemas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
    post:
      summary: Create schema
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSchema'
      responses:
        '200':
          description: Created schema
        '400':
          description: Validation error
  /api/schemas/{id}:
    get:
      summary: Get schema by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Schema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schema'
        '404':
          description: Not found
    patch:
      summary: Update schema
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSchema'
      responses:
        '200':
          description: Updated schema
        '404':
          description: Not found
    delete:
      summary: Delete schema
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found
  /api/schemas/{id}/sync-weaviate:
    post:
      summary: Manually sync schema to Weaviate
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Synced successfully
        '400':
          description: Schema must be published first
        '404':
          description: Not found
        '500':
          description: Weaviate sync failed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
    Schema:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        content: { type: string }
        systemPrompt:
          type: string
          nullable: true
          maxLength: 5000
          description: Optional system prompt for AI agents (client hint; server does not enforce maxLength)
        version: { type: integer }
        isPublished: { type: boolean }
        weaviateCollectionId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateSchema:
      type: object
      required: [name, content]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        content: { type: string }
        systemPrompt: { type: string, nullable: true, maxLength: 5000 }
        isPublished: { type: boolean, default: false }
    UpdateSchema:
      type: object
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        content: { type: string }
        systemPrompt: { type: string, nullable: true, maxLength: 5000 }
        isPublished: { type: boolean }
...